Undid nasty wrapper class, moved deferred setup into individual device submodules
