+++ library/rainbowhat/HT16K33.py
@@ -45,6 +45,7 @@ class HT16K33(object):
         self._i2c_addr = address
         self._device = i2c
         self.buffer = bytearray([0]*16)
+        self._is_setup = False
 
     def begin(self):
         """Initialize driver with LEDs enabled and all turned off."""
@@ -101,6 +102,11 @@ class HT16K33(object):
 
     def write_display(self):
         """Write display buffer to display hardware."""
+
+        if not self._is_setup:
+            self.begin()
+            self._is_setup = True
+
         for i, value in enumerate(self.buffer):
             self._device.write_byte_data(self._i2c_addr, i, value)
 
@@ -108,3 +114,4 @@ class HT16K33(object):
         """Clear contents of display buffer."""
         for i, value in enumerate(self.buffer):
             self.buffer[i] = 0
+
