--- library/rainbowhat/i2c_bus.py
@@ -1,22 +1,0 @@
-from sys import version_info
-
-try:
-    import smbus
-except ImportError:
-    if version_info[0] < 3:
-        raise ImportError("This library requires python-smbus\nInstall with: sudo apt-get install python-smbus")
-    elif version_info[0] == 3:
-        raise ImportError("This library requires python3-smbus\nInstall with: sudo apt-get install python3-smbus")
-
-try:
-    import RPi.GPIO as GPIO
-except ImportError:
-    raise ImportError("This library requires the RPi.GPIO module\nInstall with: sudo pip install RPi.GPIO")
-
-
-bus = None
-
-if GPIO.RPI_REVISION == 2 or GPIO.RPI_REVISION == 3:
-    bus = smbus.SMBus(1)
-else:
-    bus = smbus.SMBus(0)
